+++ src/features/index.js
@@ -1,4 +1,17 @@
 import { initFCM } from './notification/services/initFCM'
-import MyPageMenu from './user/components/MyPageMenu'
+import MyPage from './user/pages/MyPage'
+import HomePage from './product/pages/HomePage'
+import ProductDetailPage from './product/pages/ProductDetailPage'
+import ProductRegisterPage from './product/pages/ProductRegisterPage'
+import BiddingListPage from './product/pages/BiddingListPage'
+import SearchPage from './product/pages/ProductSearchPage'
 
-export { initFCM, MyPageMenu }
+export {
+  initFCM,
+  HomePage,
+  ProductDetailPage,
+  ProductRegisterPage,
+  BiddingListPage,
+  SearchPage,
+  MyPage,
+}
