--- src/pages/HomePage.jsx
@@ -1,9 +1,0 @@
-function HomePage() {
-  return (
-    <>
-      <h1>Home</h1>
-    </>
-  )
-}
-
-export default HomePage
